Test code(see summary for details)
#testing and experimenting. Thinking of making a jump table containing
the offset of all the 9 letter words. The 9 letter words will have the
words that can be matched with thier
#letters directly after it with a RS(30 in decimal)(record seperator)
signifying the end of the record. the program will choose a random #,
use that number in the jumptable to jump
#to the address of the 9 letter word and will use that word and its
subsequent matches to make a seperate wordlist that will be used for the
actual game.