Update
The program now prints out the words that were found. I will try and
update the find function so that it searches through the found words
list and only continues if the word hasnt been input yet. Also changed
the max letters the user can input from 10 to 11 so that 9 letter words
were recognized as valid.